Move `FallbackEngine` and `NotSupportedError` to their own package.
