+++ pkg/streamingpromql/engine_test.go
@@ -14,6 +14,8 @@ import (
 	"github.com/prometheus/prometheus/model/timestamp"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/mimir/pkg/streamingpromql/compat"
 )
 
 func TestUnsupportedPromQLFeatures(t *testing.T) {
@@ -46,13 +48,13 @@ func TestUnsupportedPromQLFeatures(t *testing.T) {
 		t.Run(expression, func(t *testing.T) {
 			qry, err := engine.NewRangeQuery(ctx, nil, nil, expression, time.Now().Add(-time.Hour), time.Now(), time.Minute)
 			require.Error(t, err)
-			require.ErrorIs(t, err, NotSupportedError{})
+			require.ErrorIs(t, err, compat.NotSupportedError{})
 			require.EqualError(t, err, "not supported by streaming engine: "+expectedError)
 			require.Nil(t, qry)
 
 			qry, err = engine.NewInstantQuery(ctx, nil, nil, expression, time.Now())
 			require.Error(t, err)
-			require.ErrorIs(t, err, NotSupportedError{})
+			require.ErrorIs(t, err, compat.NotSupportedError{})
 			require.EqualError(t, err, "not supported by streaming engine: "+expectedError)
 			require.Nil(t, qry)
 		})
@@ -69,7 +71,7 @@ func TestUnsupportedPromQLFeatures(t *testing.T) {
 		t.Run(expression, func(t *testing.T) {
 			qry, err := engine.NewInstantQuery(ctx, nil, nil, expression, time.Now())
 			require.Error(t, err)
-			require.ErrorIs(t, err, NotSupportedError{})
+			require.ErrorIs(t, err, compat.NotSupportedError{})
 			require.EqualError(t, err, "not supported by streaming engine: "+expectedError)
 			require.Nil(t, qry)
 		})
